# src/test/resources/openapi/test-api.yaml
openapi: 3.0.3
info:
  title: Spec for api tests.
  version: '1'
paths:
...
  /response/void-not-found:
    get:
      operationId: voidNotFound
      tags: [response]
      parameters:
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        204:
          description: Ok.
        404:
          description: Not found.
...
  /response/optional:
    get:
      operationId: optional
      tags: [response]
      parameters:
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                type: string
        404:
          description: Not found.
...
  /response/array-simple-not-found:
    get:
      operationId: arraySimpleNotFound
      tags: [response]
      parameters:
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        404:
          description: Not found.
...
  /response/array-complex-not-found:
    get:
      operationId: arrayComplexNotFound
      tags: [response]
      parameters:
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StringModel'
        404:
          description: Not found.
  /response/multiple-response-codes:
    get:
      operationId: multipleResponseCodes
      tags: [response]
      parameters:
        - name: redirect
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
        300:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
  /response/multiple-response-codes-not-found:
    get:
      operationId: multipleResponseCodesNotFound
      tags: [response]
      parameters:
        - name: redirect
          in: query
          required: true
          schema:
            type: boolean
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
        300:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
        404:
          description: Not found.
  /response/multiple-datatypes:
    get:
      operationId: multipleDataTypes
      tags: [response]
      parameters:
        - name: redirect
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
        300:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NumberModel'
  /response/multiple-datatypes-not-found:
    get:
      operationId: multipleDataTypesNotFound
      tags: [response]
      parameters:
        - name: redirect
          in: query
          required: true
          schema:
            type: boolean
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringModel'
        300:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NumberModel'
        404:
          description: Not found.
...
components:
  schemas:
    StringModel:
      type: object
      properties:
        foo:
          type: string
      required:
        - foo
    NumberModel:
      type: object
      properties:
        bar:
          type: number
...